requirements.txt
numpy
fastapi
uvicorn
tensorflow
opendatasets
pillow
opendatasets